+++ generator/integration-tests/basics/lib/other.dart
@@ -1,0 +1,9 @@
+import 'package:objectbox/objectbox.dart';
+
+@Entity()
+class C {
+  @Id()
+  int id;
+
+  C();
+}