+++ lib/src/bindings/structs.dart
@@ -4,6 +4,9 @@ import "package:ffi/ffi.dart" show allocate, free, Utf8;
 
 import '../common.dart';
 
+// Disable some linter rules for this file
+// ignore_for_file: camel_case_types
+
 // Note: IntPtr seems to be the the correct representation for size_t: "Represents a native pointer-sized integer in C."
 
 /// Represents the following C struct:
@@ -30,7 +33,7 @@ class OBX_id_array extends Struct {
     array.length = items.length;
     array._itemsPtr = allocate<Uint64>(count: array.length);
     for (int i = 0; i < items.length; ++i) {
-      array._itemsPtr.elementAt(i).value = items[i];
+      array._itemsPtr[i] = items[i];
     }
 
     // call the function with the structure and free afterwards
@@ -81,7 +84,7 @@ class OBX_bytes extends Struct {
     // create a copy of the data
     bytes._dataPtr = allocate<Uint8>(count: bytes.length);
     for (int i = 0; i < data.length; ++i) {
-      bytes._dataPtr.elementAt(i).value = data[i];
+      bytes._dataPtr[i] = data[i];
     }
 
     return ptr;